backend/e1/e2/requirements.txt
asgiref==3.7.2
Django==4.1.12
django-cors-headers==4.3.0
django-mongodb-engine==0.6.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.4.2
PyJWT==2.8.0
pymongo==3.12.3
python-snappy==0.6.1
pytz==2023.3.post1
sqlparse==0.2.4
typing_extensions==4.8.0
tzdata==2023.3
winkerberos==0.9.1
